test: refactor mtls test structure
