--- spring-cloud-alibaba-starters/spring-cloud-starter-mtls/src/test/java/com/alibaba/cloud/mtls/client/MtlsApplication.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2022-2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.cloud.mtls.client;
-
-import com.alibaba.cloud.governance.istio.XdsAutoConfiguration;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.http.server.reactive.SslInfo;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.server.ServerWebExchange;
-
-@SpringBootApplication(exclude = { XdsAutoConfiguration.class })
-public class MtlsApplication {
-
-	@RestController
-	static class EchoController {
-
-		@RequestMapping("/echo")
-		public String echo(ServerWebExchange exchange) {
-			if (exchange.getRequest().getSslInfo() != null) {
-				SslInfo sslInfo = exchange.getRequest().getSslInfo();
-				if (sslInfo.getPeerCertificates() != null) {
-					return "success";
-				}
-			}
-			throw new RuntimeException("No client certificate");
-		}
-
-	}
-
-}
