--- spring-cloud-alibaba-starters/spring-cloud-starter-mtls/src/test/java/com/alibaba/cloud/mtls/server/webflux/MtlsWebfluxApplication.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2022-2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.cloud.mtls.server.webflux;
-
-import com.alibaba.cloud.governance.istio.XdsAutoConfiguration;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@SpringBootApplication(exclude = { XdsAutoConfiguration.class })
-public class MtlsWebfluxApplication {
-
-	@RestController
-	static class EchoController {
-
-		@RequestMapping("/echo")
-		public String echo() {
-			return "echo";
-		}
-
-	}
-
-}
